requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
pydantic
python-jose[cryptography]
python-multipart
psycopg2-binary
pytest
httpx
